requirements.txt
Flask==2.2.3
PyYAML==6.0
requests==2.28.1
pytest==7.2.2
elasticsearch==8.0.0
requests
pandas
PyYAML